requirements.txt
# =============================================================================
# IMAS Manager - Python Dependencies
# =============================================================================

# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-environ>=0.11,<1.0

# Database
psycopg2-binary>=2.9,<3.0

# Async / Task Queue
celery[redis]>=5.3,<6.0
redis>=5.0,<6.0

# WSGI Server
gunicorn>=21.0,<23.0

# API & Serialization
drf-spectacular>=0.27,<1.0  # OpenAPI schema generation

# Google Drive Integration (Phase 2)
google-api-python-client>=2.100,<3.0
google-auth>=2.23,<3.0

# Notifications (Phase 2)
slack-sdk>=3.23,<4.0
httpx>=0.25,<1.0  # Async HTTP client for webhooks

# Development & Testing
pytest>=7.4,<9.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<6.0
factory-boy>=3.3,<4.0

# Code Quality
ruff>=0.1,<1.0
mypy>=1.6,<2.0
django-stubs>=4.2,<6.0

# Utilities
python-dateutil>=2.8,<3.0
whitenoise==6.11.0
channels==4.3.2
channels_redis==4.3.0
daphne==4.2.1
django-redis==6.0.0
